Create Example function

This creates a simulated reduced dimension dataset

create_differential_topology(
  n_cells = 200,
  noise = 0.15,
  shift = 10,
  unbalance_level = 0.9,
  speed = 1
)

Arguments

n_cells
The number of cells in the dataset.

noise
Amount of noise. Between 0 and 1.

shift
How much should the top lineage shift in condition B.

unbalance_level
How much should the bottom lineage be unbalanced toward condition A.

speed
How fast the cells from condition B should differentiate

Value

A list with two components

sd: An n_cells by 4 dataframe that contains the reduced dimensions coordinates, lineage assignment (1 or 2) and condition assignment (A or B) for each cell.

mst: a data.frame that contains the skeleton of the trajectories

Examples

sd <- create_differential_topology()