Merge slingshots datasets

If trajectory inference needs to be manually done condition per condition, this allows to merge them into one. It requires manual mapping of lineages.

merge_sds(..., mapping, condition_id = seq_len(ncol(mapping)), scale = FALSE)

Arguments

...
Slingshot datasets

mapping
a matrix, one column per dataset. Each row amounts to lineage mapping.

condition_id
A vector of condition for each condition. Default to integer values in order of appearance

scale
If TRUE (default), lineages that are mapped are scaled to have the same length.

Value

A modified slingshot dataset that can be used for downstream steps.

Details

The function assumes that each lineage in a dataset maps to exactly one lineage in another dataset. Anything else needs to be done manually.

Examples

data(list = 'slingshotExample', package = "slingshot")
if (!"cl" %in% ls()) {
  rd <- slingshotExample$rd
  cl <- slingshotExample$cl
}
sds <- slingshot::slingshot(rd, cl)

#> Using full covariance matrix

merge_sds(sds, sds, mapping = matrix(c(1, 2, 1, 2), nrow = 2))

#> class: SlingshotDataSet 
#> 
#>  Samples Dimensions
#>      280          2
#> 
#> lineages: 2 
#> Lineage1: 1  2  3  5  
#> Lineage2: 1  2  3  4  
#> 
#> curves: 2 
#> Curve1: Length: 15.045	Samples: 201.2
#> Curve2: Length: 15.126	Samples: 207