weights_from_pst

Most trajectory inference methods do not perform soft assignment but instead assign cells to all possible lineages before a branching point, and then to one or another. This function re-creates a weight matrix from those matrices of pseudotime

weights_from_pst(pseudotime, ...)

# S4 method for matrix
weights_from_pst(pseudotime)

# S4 method for data.frame
weights_from_pst(pseudotime)

Arguments

pseudotime
A matrix or data.frame of \[ncells\] by \[nCurves\].

...
Other parameters including:

Value

A object of the same type and dimensions as the original object, with the weights for each curve and cell.

Examples

data(list = 'slingshotExample', package = "slingshot")
if (!"cl" %in% ls()) {
  rd <- slingshotExample$rd
  cl <- slingshotExample$cl
}
sds <- slingshot::slingshot(rd, cl)
weights_from_pst(slingshot::slingPseudotime(sds))

#>        Lineage1 Lineage2
#>   [1,]      0.5      0.5
#>   [2,]      0.5      0.5
#>   [3,]      0.5      0.5
#>   [4,]      0.5      0.5
#>   [5,]      0.5      0.5
#>   [6,]      0.5      0.5
#>   [7,]      0.5      0.5
#>   [8,]      0.5      0.5
#>   [9,]      0.5      0.5
#>  [10,]      0.5      0.5
#>  [11,]      0.5      0.5
#>  [12,]      0.5      0.5
#>  [13,]      0.5      0.5
#>  [14,]      0.5      0.5
#>  [15,]      0.5      0.5
#>  [16,]      0.5      0.5
#>  [17,]      0.5      0.5
#>  [18,]      0.5      0.5
#>  [19,]      0.5      0.5
#>  [20,]      0.5      0.5
#>  [21,]      0.5      0.5
#>  [22,]      0.5      0.5
#>  [23,]      0.5      0.5
#>  [24,]      1.0      0.0
#>  [25,]      0.5      0.5
#>  [26,]      0.5      0.5
#>  [27,]      0.5      0.5
#>  [28,]      0.5      0.5
#>  [29,]      0.5      0.5
#>  [30,]      0.5      0.5
#>  [31,]      0.5      0.5
#>  [32,]      0.5      0.5
#>  [33,]      0.5      0.5
#>  [34,]      0.5      0.5
#>  [35,]      0.5      0.5
#>  [36,]      0.5      0.5
#>  [37,]      0.5      0.5
#>  [38,]      0.5      0.5
#>  [39,]      0.5      0.5
#>  [40,]      0.5      0.5
#>  [41,]      0.5      0.5
#>  [42,]      0.5      0.5
#>  [43,]      0.5      0.5
#>  [44,]      0.5      0.5
#>  [45,]      0.5      0.5
#>  [46,]      0.5      0.5
#>  [47,]      0.5      0.5
#>  [48,]      0.5      0.5
#>  [49,]      0.5      0.5
#>  [50,]      0.5      0.5
#>  [51,]      0.5      0.5
#>  [52,]      0.5      0.5
#>  [53,]      0.5      0.5
#>  [54,]      0.0      1.0
#>  [55,]      0.5      0.5
#>  [56,]      1.0      0.0
#>  [57,]      0.5      0.5
#>  [58,]      0.5      0.5
#>  [59,]      0.5      0.5
#>  [60,]      0.5      0.5
#>  [61,]      1.0      0.0
#>  [62,]      1.0      0.0
#>  [63,]      1.0      0.0
#>  [64,]      1.0      0.0
#>  [65,]      1.0      0.0
#>  [66,]      1.0      0.0
#>  [67,]      1.0      0.0
#>  [68,]      1.0      0.0
#>  [69,]      1.0      0.0
#>  [70,]      1.0      0.0
#>  [71,]      1.0      0.0
#>  [72,]      1.0      0.0
#>  [73,]      1.0      0.0
#>  [74,]      1.0      0.0
#>  [75,]      1.0      0.0
#>  [76,]      1.0      0.0
#>  [77,]      1.0      0.0
#>  [78,]      1.0      0.0
#>  [79,]      1.0      0.0
#>  [80,]      1.0      0.0
#>  [81,]      1.0      0.0
#>  [82,]      1.0      0.0
#>  [83,]      1.0      0.0
#>  [84,]      1.0      0.0
#>  [85,]      1.0      0.0
#>  [86,]      1.0      0.0
#>  [87,]      1.0      0.0
#>  [88,]      1.0      0.0
#>  [89,]      1.0      0.0
#>  [90,]      1.0      0.0
#>  [91,]      0.5      0.5
#>  [92,]      0.5      0.5
#>  [93,]      0.5      0.5
#>  [94,]      0.5      0.5
#>  [95,]      0.5      0.5
#>  [96,]      0.5      0.5
#>  [97,]      0.5      0.5
#>  [98,]      0.5      0.5
#>  [99,]      0.5      0.5
#> [100,]      0.5      0.5
#> [101,]      0.5      0.5
#> [102,]      0.5      0.5
#> [103,]      0.5      0.5
#> [104,]      0.0      1.0
#> [105,]      0.5      0.5
#> [106,]      0.5      0.5
#> [107,]      0.5      0.5
#> [108,]      0.5      0.5
#> [109,]      0.0      1.0
#> [110,]      0.0      1.0
#> [111,]      0.0      1.0
#> [112,]      0.0      1.0
#> [113,]      0.0      1.0
#> [114,]      0.0      1.0
#> [115,]      0.5      0.5
#> [116,]      0.0      1.0
#> [117,]      0.0      1.0
#> [118,]      0.0      1.0
#> [119,]      0.0      1.0
#> [120,]      0.0      1.0
#> [121,]      0.0      1.0
#> [122,]      0.0      1.0
#> [123,]      0.0      1.0
#> [124,]      0.0      1.0
#> [125,]      0.0      1.0
#> [126,]      0.0      1.0
#> [127,]      0.0      1.0
#> [128,]      0.0      1.0
#> [129,]      0.0      1.0
#> [130,]      0.0      1.0
#> [131,]      0.0      1.0
#> [132,]      0.0      1.0
#> [133,]      0.0      1.0
#> [134,]      0.0      1.0
#> [135,]      0.0      1.0
#> [136,]      0.0      1.0
#> [137,]      0.0      1.0
#> [138,]      0.0      1.0
#> [139,]      0.0      1.0
#> [140,]      0.0      1.0